--- a/frontend/src/main/react/src/components/Balance.tsx
+++ b/frontend/src/main/react/src/components/Balance.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
-import Typography from "@material-ui/core/Typography";
-import Grid from "@material-ui/core/Grid";
+import Typography from "@mui/material/Typography";
+import Grid from "@mui/material/Grid";
 import { SpinnerWithText } from "./SpinnerWithText";
 import { BankApiClient as bankApiClient } from "../utils/BankApiClient";
 import { withLoading } from "../utils/util";