--- a/frontend/src/main/react/src/components/Logout.tsx
+++ b/frontend/src/main/react/src/components/Logout.tsx
@@ -18,8 +18,7 @@ export const Logout = (props: { handleRefresh: () => void }) => {
         variant={"button"}
         sx={{ display: { sm: "block", xs: "none" } }}
       >
-        {" "}
-        Logout{" "}
+        Logout
       </Typography>
     </Button>
   );