--- a/frontend/src/main/react/src/components/Potters.tsx
+++ b/frontend/src/main/react/src/components/Potters.tsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from "react";
 import { SpinnerWithText } from "./SpinnerWithText";
 import { BankApiClient as bankApiClient } from "../utils/BankApiClient";
 import { withLoading } from "../utils/util";
-import Typography from "@material-ui/core/Typography";
-import { Grid } from "@material-ui/core";
-import Switch from "@material-ui/core/Switch";
+import Typography from "@mui/material/Typography";
+import { Grid } from "@mui/material";
+import Switch from "@mui/material/Switch";
 import { Potter } from "../utils/domain";
 
 interface SeasonPotters {