--- a/frontend/src/main/react/src/components/Refresh.tsx
+++ b/frontend/src/main/react/src/components/Refresh.tsx
@@ -1,13 +1,19 @@
 import React from "react";
-import RefreshIcon from "@material-ui/icons/Refresh";
-import Button from "@material-ui/core/Button";
-import Hidden from "@material-ui/core/Hidden";
+import RefreshIcon from "@mui/icons-material/Refresh";
+import Button from "@mui/material/Button";
+import Typography from "@mui/material/Typography";
 
 export const Refresh = (props: { handleRefresh: () => void }) => {
   return (
+    // TODO consider creating a XyzButton that accepts a clickhandler, icon and text to show for sm and up
     <Button variant="contained" onClick={props.handleRefresh}>
       <RefreshIcon />
-      <Hidden xsDown>Refresh</Hidden>
+      <Typography
+        variant={"button"}
+        sx={{ display: { sm: "block", xs: "none" } }}
+      >
+        Refresh
+      </Typography>
     </Button>
   );
 };