--- a/frontend/src/main/react/src/components/SpinnerWithText.tsx
+++ b/frontend/src/main/react/src/components/SpinnerWithText.tsx
@@ -1,39 +1,31 @@
 import React from "react";
-import { ClassNameMap, makeStyles } from "@mui/styles";
 import Typography from "@mui/material/Typography";
 import Grid from "@mui/material/Grid";
 import CircularProgress from "@mui/material/CircularProgress";
 
-const useStyles: (size: number) => (props?: any) => ClassNameMap = (
-  size: number
-) =>
-  makeStyles({
-    alignCenter: {
-      alignItems: "center",
-      justifyContent: "center",
-      display: "flex",
-      minHeight: size,
-    },
-  });
+const minHeight = {
+  sm: 50,
+  md: 100,
+  lg: 150,
+};
 
 export const SpinnerWithText = (props: {
   text: string;
   size?: "sm" | "md" | "lg";
 }) => {
   const styles = {
-    sm: useStyles(50),
-    md: useStyles(100),
-    lg: useStyles(150),
+    alignItems: "center",
+    justifyContent: "center",
+    display: "flex",
+    minHeight: minHeight[props.size || "md"],
   };
-  const { size = "md" } = props;
-  const classes = (styles[size] || styles["md"])();
   return (
     <>
       <Grid container>
-        <Grid item xs={12} className={classes.alignCenter}>
+        <Grid item xs={12} sx={styles}>
           <CircularProgress />
         </Grid>
-        <Grid item xs={12} className={classes.alignCenter}>
+        <Grid item xs={12} sx={styles}>
           <Typography variant="h6">{props.text}</Typography>
         </Grid>
       </Grid>