--- a/frontend/src/main/react/src/components/TopUp.tsx
+++ b/frontend/src/main/react/src/components/TopUp.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { Button } from "@material-ui/core";
-import Grid from "@material-ui/core/Grid";
-import Typography from "@material-ui/core/Typography";
+import { Button } from "@mui/material";
+import Grid from "@mui/material/Grid";
+import Typography from "@mui/material/Typography";
 
 const bunqMeUrl = "https://bunq.me/tovoheren5";