--- a/frontend/src/main/react/src/components/Transactions.tsx
+++ b/frontend/src/main/react/src/components/Transactions.tsx
@@ -1,22 +1,21 @@
 import React, { useEffect, useState } from "react";
-import Table from "@material-ui/core/Table";
-import TableBody from "@material-ui/core/TableBody";
-import TableCell from "@material-ui/core/TableCell";
-import TableContainer from "@material-ui/core/TableContainer";
-import TableHead from "@material-ui/core/TableHead";
-import TableRow from "@material-ui/core/TableRow";
-import Paper from "@material-ui/core/Paper";
+import Table from "@mui/material/Table";
+import TableBody from "@mui/material/TableBody";
+import TableCell from "@mui/material/TableCell";
+import TableContainer from "@mui/material/TableContainer";
+import TableHead from "@mui/material/TableHead";
+import TableRow from "@mui/material/TableRow";
+import Paper from "@mui/material/Paper";
 import { SpinnerWithText } from "./SpinnerWithText";
 import { BankApiClient as bankApiClient } from "../utils/BankApiClient";
 import { formattedDate, formattedTime, withLoading } from "../utils/util";
+import { createStyles, makeStyles } from "@mui/styles";
 import {
-  createStyles,
-  makeStyles,
   TableFooter,
   TablePagination,
   useMediaQuery,
   useTheme,
-} from "@material-ui/core";
+} from "@mui/material";
 import { Transaction } from "../utils/domain";
 
 const useStyles = makeStyles(() =>