… frontend/package.json
… frontend/src/main/react/src/components/Users.tsx
… frontend/src/main/react/src/components/events/EditableTextField.tsx
… frontend/src/main/react/src/components/events/EventDetails.tsx
… frontend/src/main/react/src/components/events/EventForm.tsx
… frontend/src/main/react/src/components/events/Events.tsx
… frontend/src/main/react/src/components/events/EventsList.tsx
… frontend/src/main/react/src/components/events/EventsTable.tsx
… frontend/src/main/react/src/components/events/SelectedUser.tsx
… frontend/src/main/react/src/hooks/alertsHook.tsx
… frontend/src/main/react/src/views/EventsPage.tsx
… frontend/src/main/react/src/views/Login.tsx
… frontend/src/main/react/src/views/Overview.tsx
… frontend/src/main/react/src/views/TransactionsPage.tsx
… frontend/src/main/react/src/views/UsersPage.tsx
… frontend/webpack.config.js