- Add themeprovider to app
- Add MUI 5 dependencies
  - included @mui/x-date-pickers (as replacement for @material-ui/pickers)
  - included emotion for typesafe css
- Remove MUI 4 dependencies, and fix breaking component changes

  Follows: https://mui.com/material-ui/migration/v5-component-changes/

  Noteworthy:
  - Fixes in createStyles and withStyles
  - Deprecated `<Hidden>` component (mostly move to:

```tsx
<Typography variant={"button"} sx={{ display: { sm: "block", xs: "none" } }}>
  Terug naar de veiligheid
</Typography>
```
  - App is wrapped in <StyledEngineProvider injectFirst> to support JSS (prepare for migration to emotion
  - Shit ton of renames from @material-ui/* to @mui/*
  - Verwijderen van AttendeeStyledButton, en fallback op normale Button. 'uncertain' is mapped naar 'warning', en die kleur is aangepast in de palette van de theme.

Apply webpack optimisations to reduce build times (dev: 30s to 10s, prod 60s to 10s)

Replace JSS with emotion

Remove @mui/styles

Upgrade notistack to v2